requirements.txt
customtkinter==5.2.0
mysql_connector_repackaged==0.3.1
